checkout flash request

+++ app/Http/Controllers/BasketController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Order;
+use App\Models\Product;
 use Illuminate\Http\Request;
 
 class BasketController extends Controller
@@ -12,11 +13,10 @@ class BasketController extends Controller
         if(!is_null($orderId)) {
             $order = Order::findOrFail($orderId);
         }
-
         return view('pages.basket', compact('order'));
     }
 
-    public function basketAdd($productId, Request $request) {
+    public function basketAdd($productId) {
         $orderId = session('orderId');
         if(is_null($orderId)) {
             $order = Order::create()->id;
@@ -24,17 +24,70 @@ class BasketController extends Controller
         } else {
             $order = Order::find($orderId);
         }
-        $order->products()->attach($productId);
-        return view('pages.basket', compact('order'));
+
+        if($order->products->contains($productId)) {
+            $pivotRow = $order->products()->where('product_id', $productId)->first()->pivot;
+            $pivotRow->count++;
+            $pivotRow->update();
+        } else {
+            $order->products()->attach($productId);
+        }
+        $product = Product::find($productId);
+        if($product) {
+            session()->flash('success', 'Product "'.$product->name.'" has been added to basket successfully!');
+        } else {
+            session()->flash('warning', 'Error 404');
+        }
+
+        return redirect()->route('basket');
     }
 
     public function basketRemove($productId) {
         $orderId = session('orderId');
         if(is_null($orderId)) {
-            return view('pages.basket', compact('order'));
+            return redirect()->route('basket');
         }
         $order = Order::find($orderId);
-        $order->products()->detach($productId);
-        return view('pages.basket', compact('order'));
+
+        if($order->products->contains($productId)) {
+            $pivotRow = $order->products()->where('product_id', $productId)->first()->pivot;
+            if($pivotRow->count < 2) {
+                $order->products()->detach($productId);
+            } else {
+                $pivotRow->count--;
+                $pivotRow->update();
+            }
+        }
+        $product = Product::find($productId);
+        if($product) {
+            session()->flash('warning', 'Product "'.$product->name.'" has been deleted from basket');
+        }
+
+        return redirect()->route('basket');
+    }
+
+    public function basketCheckout() {
+        $orderId = session('orderId');
+        if(is_null($orderId)) {
+            return redirect()->route('home');
+        }
+        $order = Order::find($orderId);
+        return view('pages.checkout', compact('order'));
+    }
+
+    public function basketConfirm(Request $request) {
+        $orderId = session('orderId');
+        if(is_null($orderId)) {
+            return redirect()->route('home');
+        }
+        $order = Order::find($orderId);
+        $success = $order->saveOrder($request->name, $request->phone);
+
+        if($success) {
+            session()->flash('success', 'Your Order #'.$order->id.' has been created successfully!');
+        } else {
+            session()->flash('warning', 'Error 404');
+        }
+        return redirect()->route('home');
     }
 }
